+++ application/controllers/User.php
@@ -8,11 +8,6 @@ class User extends CI_Controller
 	{
 		parent::__construct();
 		test_login(2);
-		$config['upload_path'] = 'assets/profileimages';
-		$config['allowed_types'] = 'gif|jpg|png';
-		$config['encrypt_name'] = TRUE;
-		$this->load->library('upload', $config);
-		$this->load->library('image_lib');
 	}
 
 	/**
@@ -180,6 +175,11 @@ class User extends CI_Controller
 	}
 
 	public function profile() {
+		$config['upload_path'] = 'assets/profileimages';
+		$config['allowed_types'] = 'gif|jpg|png';
+		$config['encrypt_name'] = TRUE;
+		$this->load->library('upload', $config);
+		$this->load->library('image_lib');
 
 		$this->form_validation->set_rules('updateFirstName', 'First Name', 'trim|max_length[50]|min_length[1]|required');
 		$this->form_validation->set_rules('updateLastName', 'Last Name', 'trim|max_length[50]|min_length[1]|required');
@@ -224,7 +224,9 @@ class User extends CI_Controller
 					$this->image_lib->initialize($configer);
 					$this->image_lib->resize();
 
-					unlink('assets/profileimages/'.$_SESSION['user']['profilepath']);		// delete the old image
+					if($_SESSION['user']['profilepath'] != 'default.jpg') {
+						unlink('assets/profileimages/'.$_SESSION['user']['profilepath']);		// delete the old image
+					}
 
 					$data['profilepath'] = $image_data['file_name'];
